And/feat#364 group mapping

+++ android/data/src/main/java/com/sixkids/data/model/request/ChallengeCreateRequest.kt
@@ -15,6 +15,6 @@ data class ChallengeCreateRequest(
 
 data class GroupRequest(
     val headCount: Int,
-    val leaderId: Int,
+    val leaderId: Long,
     val students: List<Long>
 )
